Move classes files outside their folders
