+++ data/DB/labels.json
@@ -1,7 +1,12 @@
 [
   {
-    "id": 452,
+    "id": 129,
     "title": "y",
     "color": "c"
+  },
+  {
+    "id": 875,
+    "title": "g",
+    "color": "y"
   }
 ]
