+++ src/classes/label.rb
@@ -1,4 +1,4 @@
-require_relative '../item/item'
+require_relative './item'
 
 class Label
   attr_accessor :title, :color
