+++ src/classes/music_album.rb
@@ -1,4 +1,4 @@
-require_relative '../item/item'
+require_relative './item'
 
 class MusicAlbum < Item
   attr_accessor :on_spotify
